scripts/requirements-embeddings.txt
torch>=2.0.0
transformers>=4.40.0
umap-learn>=0.5.6
rank-bm25>=0.2.2
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
